Split ConnectionStatusContext from ConnectionStatusProvider
